requirements.txt
numpy
pandas
matplotlib
seaborn
mlflow
scikit-learn
pyYAML
types-PyYAML
Flask
joblib
tqdm
dvc
python-dotenv
python-box
SQLAlchemy
pymysql
xgboost
dagshub
#-e .